--- client/src/pages/Home-backup.tsx
@@ -1,1 +1,0 @@
-// Backup do arquivo Home.tsx antes da correção
