+++ client/src/pages/Workout.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useState, useEffect } from "react";
 import { Card, CardContent } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Badge } from "@/components/ui/badge";
@@ -71,20 +71,6 @@ export default function Workout() {
   const [showSetFeedback, setShowSetFeedback] = useState(false);
   const { showWorkoutSuccess } = useGlobalNotification();
 
-  // Refs para auto-scroll
-  const executionCardRef = useRef<HTMLDivElement>(null);
-  const timerCardRef = useRef<HTMLDivElement>(null);
-
-  // Função para scroll suave para cards importantes
-  const scrollToCard = (ref: React.RefObject<HTMLDivElement>) => {
-    if (ref.current) {
-      ref.current.scrollIntoView({
-        behavior: 'smooth',
-        block: 'center'
-      });
-    }
-  };
-
   // Scroll to top when component mounts
   useEffect(() => {
     window.scrollTo({ top: 0, behavior: 'smooth' });
@@ -118,8 +104,6 @@ export default function Workout() {
       setIsResting(false);
       setIsTimerRunning(false);
       setShowSetFeedback(false);
-      // Auto-scroll para o card de execução
-      setTimeout(() => scrollToCard(executionCardRef), 300);
     }
   };
 
@@ -138,8 +122,6 @@ export default function Workout() {
         setIsResting(true);
         setRestTime(exercise.restBetweenSeries || 90);
         setIsTimerRunning(true);
-        // Auto-scroll para o card do timer
-        setTimeout(() => scrollToCard(timerCardRef), 300);
       } else {
         // Exercise completed
         setCompletedExercises(prev => {
@@ -158,8 +140,6 @@ export default function Workout() {
     setIsResting(false);
     setIsTimerRunning(false);
     setShowSetFeedback(true);
-    // Auto-scroll para o card de execução
-    setTimeout(() => scrollToCard(executionCardRef), 300);
   };
 
   const finishIndividualWorkout = async () => {
@@ -315,18 +295,15 @@ export default function Workout() {
         <div>
           <h1 className="text-2xl font-bold mb-2">Treinos Programados</h1>
           <p className="text-muted-foreground">
-            Gere seu primeiro treino personalizado usando nossa IA
+            Nenhum treino encontrado. Sincronize com a IA para gerar treinos personalizados.
           </p>
         </div>
         <Card>
           <CardContent className="p-6 text-center">
             <p className="text-gray-500 mb-4">Você ainda não possui treinos programados.</p>
-            <p className="text-sm text-muted-foreground mb-4">
-              Vá para a tela inicial e clique em "Gerar treino" para começar
-            </p>
-            <Link href="/">
+            <Link href="/profile">
               <Button>
-                Ir para Tela Inicial
+                Ir para Perfil e Sincronizar
               </Button>
             </Link>
           </CardContent>
@@ -411,7 +388,7 @@ export default function Workout() {
 
       {/* Active Exercise Timer */}
       {activeExercise && (
-        <Card ref={executionCardRef} className="bg-primary dark:bg-primary border-0">
+        <Card className="bg-primary dark:bg-primary border-0">
           <CardContent className="p-6 text-primary-foreground dark:text-primary-foreground">
             <div className="text-center">
               <h2 className="text-lg font-semibold mb-2">
@@ -505,7 +482,7 @@ export default function Workout() {
 
       {/* Timer de Descanso Integrado - Posicionado após exercício ativo */}
       {isResting && activeExercise && (
-        <Card ref={timerCardRef} className="bg-orange-500 dark:bg-orange-600 border-0 animate-in slide-in-from-top duration-300">
+        <Card className="bg-orange-500 dark:bg-orange-600 border-0 animate-in slide-in-from-top duration-300">
           <CardContent className="py-4">
             <div className="flex items-center justify-between">
               <div className="flex items-center space-x-3">
